Adapted house management

+++ wflowapp/lib/main/actions/viewhouse/charts/Consume.dart
@@ -1,6 +1,6 @@
 class Consume {
   final int x;
-  final int y;
+  final double y;
   final bool predicted;
 
   Consume({required this.x, required this.y, required this.predicted});
